fix: Fix clean path leaking problem (#289).
Fix the problem when there is coincidence points on the global path. When 2 neighbour points coincide, they cannot be pruned and thus teb algorithm cannot make the robot move on.

Take a minimum step for 2 points to decide the direction of the robot, if the 2 neighbour points are too close(< min_prune_step), then find another point as the start point to make they are farther than min_prune_step.